重新调整所有文件 修改数据库（数据库被黑了） 完成文件基本调整 测试未完成

+++ src/main/java/cn/edu/imufe/pojo/AnswerHistoryPojo.java
@@ -1,13 +1,13 @@
 package cn.edu.imufe.pojo;
 
 public class AnswerHistoryPojo {
-	private Integer answerId;
+	private Long answerId;
 	private String title;
 	private Integer status;
-	public Integer getAnswerId() {
+	public Long getAnswerId() {
 		return answerId;
 	}
-	public void setAnswerId(Integer answerId) {
+	public void setAnswerId(Long answerId) {
 		this.answerId = answerId;
 	}
 	public String getTitle() {
