+++ src/main/java/cn/edu/imufe/service/AnswerService.java
@@ -1,18 +1,18 @@
 package cn.edu.imufe.service;
 import java.util.List;
 
-import cn.edu.imufe.entity.*;
+import cn.edu.imufe.po.*;
 import cn.edu.imufe.pojo.AnswerPojo;
 public interface AnswerService {
-	Answer selectByPrimaryKey(Integer id);
+	Answer selectByPrimaryKey(Long id);
 	
-	List<Integer> selectAllId();
+	List<Long> selectAllId();
 	
 	List<AnswerPojo> selectAllIdWithTitle();
 	
 	int insertSelective(Answer record);
 	
-	int deleteByPrimaryKey(Integer id);
+	int deleteByPrimaryKey(Long id);
 	
 	int updateByPrimaryKeySelective(Answer record);
 }
